Move callstack and debug to core library (incomplete)
