+++ source/library/core/private/debug.windows.h
