+++ source/library/core/private/debug.windows.rc
