+++ source/library/core/public/potato/core/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(core PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/_export.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/callstack.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/debug.h
+)
