+++ source/library/foundation/public/potato/foundation/assertion.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "debug.h"
+#include "potato/core/debug.h"
 #include "platform.h"
 #include "string_format.h"
 #include "fixed_string_writer.h"
